Supporting member now links to CASS Members table
The link is to the membership table at /about/#current-cass-members-and-leadership.

This is the result of a discussion with Lois.  We decided that having SSO-specific product catalogs was more or less redundant with area-specific product catalogs and less useful.  Linking to the actual SSO websites would unexpectedly jump people away from the CASS site, which is probably not desirable.  This seemed like a reasonable compromise that keep people on the site, but allows them to drill into the organizational stuff if they want.