Trying to make page layouts more consistent
Splash eliminates the left margin.  Single (the basis for collection pages) does not.

Right now, we're using splash where we can, but Impacts, Software, and Working Group pages are using collections/single.  Software has a left sidebar.  But Impacts and Working Groups are not currently using a left sidebar.  Are there good uses for this space?  Or do we need to try to make a "wider" class?