Revisions to Viskores PR
@vincentbolea thanks for the PR to handle the renaming of VTK-m to Viskores.  I made a few changes:

In the highlights, the `software_mentioned` field is not the place to mention the name change.

Also in the wdmapp highlight the name change was mentioned twice in the same paragraph.  It is only needed once.

In the Viskores page itself, I added a sentence at the end of the `long_description` to note the name change -- it seems important to have it here.

And in the `additional_resource_links` in on the Viskores page, I removed the E4S and Spack entries, since they're covered separately and you've already made the appropriate changes there. (When the product name changes in E4S, you'll need to make another change.)