Support swe-cse-bof events
The swe-cse-bof site makes use of headers to prettify their pages.  That doesn't work for our events, so we remove support for headers from our event layout.

The `agenda-lightning-talks` include is necessary to process these events, but the results are not displayed on the CASS site (overridden by the `description` field).

Also, add/change some icons.