Add support for `format` key in artifacts
Events on the IDEAS site used a `format` key which was rendered as "(format)".  For the CASS site, we generalized it to `note` which is rendered without additional decoration so the content creator can do what they want with it.

But since we want to share events with the IDEAS site, we need to support `format` as well.